Removed `Index` namespace and moved all index classes to main namespace
