+++ src/Bonsai.ML.Torch/BooleanIndex.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reactive.Linq;
 using TorchSharp;
 
-namespace Bonsai.ML.Torch.Index;
+namespace Bonsai.ML.Torch;
 
 /// <summary>
 /// Represents a boolean index that can be used to select elements from a tensor.
